+++ src/main/java/com/vonage/client/verify2/VerificationRequest.java
@@ -53,8 +53,8 @@ public class VerificationRequest implements Jsonable {
 		if (brand.length() > 16) {
 			throw new IllegalArgumentException("Brand cannot exceed 16 characters in length.");
 		}
-		if ((channelTimeout = builder.timeout) != null && (channelTimeout < 60 || channelTimeout > 900)) {
-			throw new IllegalArgumentException("Delivery wait timeout must be between 60 and 900 seconds.");
+		if ((channelTimeout = builder.timeout) != null && (channelTimeout < 15 || channelTimeout > 900)) {
+			throw new IllegalArgumentException("Delivery wait timeout must be between 15 and 900 seconds.");
 		}
 		if ((workflows = builder.workflows).isEmpty()) {
 			throw new IllegalStateException("At least one workflow must be defined.");